Move relevant parts of mod.rs to `sys` subcommands
